Move main to a cmd package
