+++ .gitignore
@@ -16,6 +16,7 @@ tmp/
 
 # Test binary, built with `go test -c`
 *.test
+/coverage.txt
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
