+++ config.go
@@ -1,4 +1,4 @@
-package main
+package waitfor
 
 import (
 	"errors"
