+++ test/Dockerfile
@@ -9,7 +9,7 @@ RUN go mod download
 RUN go install github.com/cucumber/godog/cmd/godog@v0.12.4
 
 ADD . .
-RUN go build -o wait-for
+RUN go build -o ./cmd/wait-for
 
 WORKDIR /app/test/
 ENTRYPOINT . files/godog.sh
